Update loop vars in CreateEntAreaRefs to be size_t to remove warning in x64 since commit: 71054a47b24586631c185292842835c89ff8be77

+++ tools/LevelBuilder/LvlProcess.cpp
@@ -502,7 +502,7 @@ void LvlBuilder::AddAreaRefs_r(core::Array<int32_t>& areaList, const Sphere& sph
 
 bool LvlBuilder::CreateEntAreaRefs(LvlEntity& worldEnt)
 {
-	int32_t i, numEnts;
+	size_t i, numEnts;
 
 	// we go throught each ent, and work out what area's it is in.
 	// each ent is then added to the entRefts set.
